requirements.txt
click==7.1.2
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
imageio==2.8.0
numpy==1.18.4
Pillow==7.1.2
